hammerhead: Fix apps2d to usb storage
While this may not be a very good idea to use (do you really want
to have a usb stick sticking out of your phone??), the current
fstab is incorrect and needs to be fixed.

Removable storage is always noemulatedsd (and defined as such
in storage_list.xml).  The fstab should be consistent with the
rest of android.

Change-Id: I03d8e2af810bc4291a2ed6adde79a9aadbc7a755
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>